Allow generation of source-code FMUs
All generated (modelica-exchange) FMUs now include the source code of
the runtime, so that it is possible to compile an (almost) static
FMU.

By default, the runtime is still linked to the generated FMU, but
this can be changed by using -d=buildStaticSourceFMU.

The makefiles for the FMUs have been changed, such that we first
generate a source-code FMU, copying in all the files. Then a second
Makefile is used to compile the binary. Both make commands are
executed by default for translateModelFMU, but this API function will
only generate pure source-code FMUs in the future (buildModelFMU will
also compile the FMU, and there should be an API to compile
additional targets for an existing FMU).

FMUs no longer use garbage collection by default since some
simulators used libgc, and it just caused all sorts of headaches.

+++ SimulationRuntime/c/simulation/simulation_info_xml.c
@@ -164,7 +164,7 @@ static void XMLCALL startElement(void *voidData, const char *name, const char **
       userData->xml->equationInfo[userData->curIndex].parent = 0;
     }
     if(userData->curIndex != ix) {
-      throwStreamPrint(NULL, "%s: Info XML %s got equation with index %ld, expected %ld", __FILE__, userData->xml->fileName, ix, userData->curIndex);
+      throwStreamPrint(NULL, "%s: Info XML %s got equation with index %ld, expected %ld", __FILE__, userData->xml->fileName, (long) ix, (long) userData->curIndex);
     }
     userData->xml->equationInfo[userData->curIndex].id = userData->curIndex;
     userData->xml->equationInfo[userData->curIndex].profileBlockIndex = measure_time_flag & 2 ? userData->curIndex : -1; /* TODO: Set when parsing other tags */
@@ -290,7 +290,7 @@ void modelInfoXmlInit(MODEL_DATA_XML* xml)
         mmc_uint_t line = XML_GetCurrentLineNumber(parser);
         fclose(file);
         XML_ParserFree(parser);
-        throwStreamPrint(NULL, "%s: Error: failed to read the XML file %s: %s at line %lu", __FILE__, xml->fileName, err, line);
+        throwStreamPrint(NULL, "%s: Error: failed to read the XML file %s: %s at line %lu", __FILE__, xml->fileName, err, (unsigned long) line);
       }
     } while(!done);
     fclose(file);
@@ -299,7 +299,7 @@ void modelInfoXmlInit(MODEL_DATA_XML* xml)
       const char *err = XML_ErrorString(XML_GetErrorCode(parser));
       mmc_uint_t line = XML_GetCurrentLineNumber(parser);
       XML_ParserFree(parser);
-      throwStreamPrint(NULL, "%s: Error: failed to read the XML data %s: %s at line %lu", __FILE__, xml->infoXMLData, err, line);
+      throwStreamPrint(NULL, "%s: Error: failed to read the XML data %s: %s at line %lu", __FILE__, xml->infoXMLData, err, (unsigned long) line);
     }
   }
   assert(xml->nEquations == userData.curIndex);
